requirements.txt
requests==2.26.0
django-environ==0.8.1
django_debug_toolbar==3.2.4
dj_database_url==0.5.0
Django==3.2.8
geopy==2.2.0
django_phonenumber_field==6.0.0
djangorestframework==3.13.1
django-admin-sortable2==1.0.3
django-extensions==3.1.5
phonenumbers==8.12.39
Pillow==8.4.0
geocoder==1.38.1
ipython==7.30.1
django-debug-toolbar~=3.2.4
rich==10.16.1
django-lifecycle==0.9.3
django-enumfields==2.1.1

flake8==4.0.1
flake8-2020==1.6.0
flake8-blind-except==0.2.0
flake8-bugbear==21.9.2
flake8-builtins==1.5.3
flake8-commas==2.1.0
flake8-comprehensions==3.7.0
flake8-debugger==4.0.0
flake8-docstrings==1.6.0
flake8-eradicate==1.2.0
flake8-polyfill==1.0.2
flake8-print==4.0.0
flake8-quotes==3.3.1
flake8-string-format==0.3.0
flake8-fixme==1.1.1
flake8-annotations-complexity==0.0.6
flake8-variables-names==0.0.4
flake8-class-attributes-order==0.1.2
flake8-broken-line==0.4.0
flake8-tidy-imports==4.5.0
flake8-typing-imports==1.11.0
dlint==0.11.0
flake8-if-statements==0.1.0
flake8-functions==0.0.6
flake8-annotations-coverage==0.0.5
flake8-expression-complexity==0.0.9
flake8-printf-formatting==1.1.2
flake8-multiline-containers==0.0.18
flake8-absolute-import==1.0
flake8-simplify==0.14.2
scspell3k==2.2
flake8-functions-names==0.0.6
flake8-obey-import-goat==0.0.1